# Use crossbeam-channel for massive speedup