--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,7 @@ bench = []
 bidirectional = ["readwrite"]
 
 [dependencies]
+crossbeam-channel = "0.3.8"
 readwrite = {version = "0.1", optional = true}
 
 [dev-dependencies]