Crossbeam-channel is apparently a lot better here.

**This introduces a breaking change**: the type of `into_inner` for `PipeReader`/`PipeWriter` is now crossbeam-channel's Sender/Reciever.

Also move benchmarks to criterion, to get some statistics on benchmarking, and to compare against `os_pipe` crate

## Performance results
```
pipe_send/pipe-rs/4096  time:   [582.11 us 592.09 us 603.99 us]
                        thrpt:  [1.6168 GiB/s 1.6493 GiB/s 1.6776 GiB/s]
                 change:
                        time:   [-94.119% -93.959% -93.763%] (p = 0.00 < 0.05)
                        thrpt:  [+1503.3% +1555.3% +1600.3%]
                        Performance has improved.
pipe_send/pipe-rs/8192  time:   [408.88 us 415.31 us 422.74 us]
                        thrpt:  [2.3101 GiB/s 2.3514 GiB/s 2.3884 GiB/s]
                 change:
                        time:   [-92.025% -91.820% -91.583%] (p = 0.00 < 0.05)
                        thrpt:  [+1088.1% +1122.5% +1153.9%]
                        Performance has improved.
pipe_send/pipe-rs/16384 time:   [332.98 us 341.16 us 349.94 us]
                        thrpt:  [2.7906 GiB/s 2.8625 GiB/s 2.9328 GiB/s]
                 change:
                        time:   [-87.484% -87.172% -86.865%] (p = 0.00 < 0.05)
                        thrpt:  [+661.32% +679.54% +698.97%]
                        Performance has improved.
pipe_send/pipe-rs/32768 time:   [291.44 us 296.80 us 303.01 us]
                        thrpt:  [3.2229 GiB/s 3.2903 GiB/s 3.3508 GiB/s]
                 change:
                        time:   [-79.435% -78.663% -77.785%] (p = 0.00 < 0.05)
                        thrpt:  [+350.14% +368.66% +386.27%]
                        Performance has improved.
pipe_send/pipe-rs/65536 time:   [282.22 us 288.76 us 295.90 us]
                        thrpt:  [3.3003 GiB/s 3.3819 GiB/s 3.4603 GiB/s]
                 change:
                        time:   [-63.272% -62.209% -60.964%] (p = 0.00 < 0.05)
                        thrpt:  [+156.17% +164.61% +172.27%]
                        Performance has improved.
```