@Dr-Emann 2019-07-18
Also, because crossbeam channel is mpmc, we could implement Clone for PipeReader as well.