@arcnmx 2019-07-18
Fancy and a pretty low-impact dependency, thanks!

Regarding cloned readers I'm a bit torn I think. If you're processing data with multiple receiving sides you'd have to be very careful about packeting your reads and writes properly, and at that point it's basically just datagrams and you should probably be using channels directly instead?

On the other hand, that same argument also applies to the writer being Clone and it already is, the type allows it so no reason not to. If anything it allows you to be lazy and clone instead of worry about Sending properly I guess.